Add M382/383 codes for 3cc syringe extension/retraction
As with M380/381 for solenoid (de)activation, it's probably a good idea
to have functions defined for syringe mechanics to avoid issues with
potential pin changes down the line.

Here, `M382` and `M382 T1` will extend the syringe for the active or
specified tool, respectively; same behavior with `M383` and `M383 T1`
for retraction.