Decouple pressure multiplier from extrusion multiplier
I had previously set this up to allow on-the-fly pressure adjustment via
the slider in OctoPrint by just piggy-backing on M221 and allowing a
toggle of this behavior via M222... this works, but leads to some
unnecessary complexity when dealing with an auger cartridge (for which a
change to the extrusion multiplier would ideally not affect pressure).

So, this repurposes M222 to handle the pressure multiplier in a pretty
much identical fashion to M221 (eg. M222 T0 S105).